Committing solution to the question I received for FB Telephone interview
Reflecting on where it went wrong, I had the right idea to traverse all left nodes first and add to stack, but for some reason the idea of doing this within an inner while loop didn't come to mind. After having half a day to reflect, it felt like I had an off day. Whilst I was calm and not showing signs of nerves, my brain just wasn't reaching for the solution, despite conceptually understanding what needed to be done, and explaining my understanding of the BST, the stack and the in order traversal. Slightly gutting because this is a question I should have aced, and may have aced with just one more day to do a few more practice exercises.

But that's what happens sometimes and you just have to take the loss, go back to the books and study smarter ready for next time. It isn't impossible that I'll get invited to the next round because I did identify the solution somewhat, but I didn't fully implement the code, so I'd still be surprised if I went any further in the process.